% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-tibble-style.R
\name{along_index}
\alias{along_index}
\title{Make or reshape an idx_tbl by arranging values along an index}
\usage{
along_index(x, idx, ...)
}
\arguments{
\item{x}{values (or idx_tbl)}

\item{idx}{index data, e.g., unique combinations of columns that will correspond to idx_cols}
}
\description{
Make or reshape an idx_tbl by arranging values along an index
}
\keyword{internal}
