% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-ops.R
\name{coalesce_by_idx}
\alias{coalesce_by_idx}
\title{Coalesce indexvectors by index}
\usage{
coalesce_by_idx(...)
}
\description{
Coalesce indexvectors by index
}
\examples{
t1 <- new_idx_tibble(tibble(year = 2005, value = 1), idx = "year")
t2 <- new_idx_tibble(tibble(year = 2010, value = 2), idx = "year")
t3 <- new_idx_tibble(tibble(year = c(2010, 2011), value = c(1.1, 1.2)), idx = "year")
coalesce_by_idx(t1, t2)
coalesce_by_idx(t1, t2, t3)
}
