% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-ops.R
\name{full_common_index}
\alias{full_common_index}
\title{Combine indices for coalesce}
\usage{
full_common_index(idx_x, idx_y)
}
\arguments{
\item{idx_x}{an index (data frame of unique combinations)}

\item{idx_y}{an index (data frame of unique combinations)}
}
\description{
Combine indices for coalesce
}
\examples{
t1 <- new_idx_tibble(tibble::tibble(year = 2005, value = 1), idx = "year")
t2 <- new_idx_tibble(tibble::tibble(year = 2010, value = 2), idx = "year")
\dontrun{
full_common_index(index(t1), index(t2))
}
}
