% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{idx_extract}
\alias{idx_extract}
\title{Extract from an idx_tbl}
\usage{
idx_extract(x, ...)
}
\arguments{
\item{x}{idx_tbl}

\item{...}{index values to extract}
}
\description{
Always returns an idx_tbl with at least 1 less index dimension than the input, analagous to \code{[[} with a nested list.
}
\examples{
pop <- idx_tibble(tidyr::population)
pop_grow <- pop / idx_extract(pop, year=2000)
}
