% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{idx_slice}
\alias{idx_slice}
\title{Slice idx_tbl by index values}
\usage{
idx_slice(x, ..., .keep_all = FALSE)
}
\arguments{
\item{x}{idx_tbl indexvector}

\item{...}{index values to slice. If unnamed, refer to idx_cols of x in order. If named they all must be named, and refer to idx_cols by name.}

\item{.keep_all}{unused, but perhaps to be used in future.}
}
\value{
idx_tbl with the same axes as input.
}
\description{
Analagous to a left-join of the input against a differently shaped index.
}
\examples{
pop <- idx_tibble(tidyr::population, idx_cols = c("country", "year"))
idx_slice(pop, "Armenia")
idx_slice(pop, year=2000)
pop_grow <- pop / idx_slice(pop, year=2000)
}
