% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-tibble-style.R
\name{index}
\alias{index}
\title{Index of an object}
\usage{
index(x, ...)
}
\arguments{
\item{x}{a idx_tbl, double, or other object compatible with indexvctrs ops}

\item{...}{unused}
}
\description{
Index of an object
}
