% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-tibble-style.R
\name{new_idx_tibble}
\alias{new_idx_tibble}
\title{Tibble-style index vector}
\usage{
new_idx_tibble(x, idx_cols)
}
\arguments{
\item{x}{tibble data}

\item{idx_cols}{character vector designating idx columns}
}
\description{
Tibble-style index vector
}
