% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-rcrd-style.R
\name{new_idx_vctr}
\alias{new_idx_vctr}
\title{Index Vector}
\usage{
new_idx_vctr(df, idx)
}
\arguments{
\item{df}{data frame with indices and \code{value} column}

\item{idx}{character vector of index columns}
}
\description{
Index Vector
}
