% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{value_at_idx}
\alias{value_at_idx}
\title{Get values of an idx_tbl for specific index values}
\usage{
value_at_idx(x, idx)
}
\arguments{
\item{x}{an idx_tbl}

\item{idx}{an index that can be used with x}
}
\description{
Get values of an idx_tbl for specific index values
}
