% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-tibble-style.R
\name{idx_tibble}
\alias{idx_tibble}
\title{Tibble-style index vector}
\usage{
idx_tibble(x, idx_cols = NULL, value_col = NULL)
}
\arguments{
\item{x}{data coercible to tbl_df}

\item{idx_cols}{character vector, name of index columns}

\item{value_col}{length-1 character vector, name of value column, defaults as "value" if present, will be renamed as \code{value}}
}
\description{
Tibble-style index vector
}
