% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-tibble-style.R
\name{new_idx_tibble}
\alias{new_idx_tibble}
\title{Tibble-style index vector}
\usage{
new_idx_tibble(tbl, idx)
}
\arguments{
\item{tbl}{on which to base it}

\item{idx}{character vector designating idx columns}
}
\description{
Tibble-style index vector
}
